+++ .gitignore
@@ -25,3 +25,4 @@ yarn-error.log*
 package-lock.json
 mereged.txt
 yarn.lock
+merged.txt
