+++ src/config/formFieldsConfig.js
@@ -1,0 +1,8 @@
+const formFieldsConfig = [
+  { name: "departureTime", defaultValue: "11:00" },
+  { name: "drivingTime", defaultValue: "45" },
+  { name: "arriveEarly", defaultValue: "30" },
+  { name: "snackTime", defaultValue: "5" },
+];
+
+export default formFieldsConfig;
